+++ enro-core/src/main/java/dev/enro/destination/fragment/container/FragmentNavigationContainerProperty.kt
@@ -4,11 +4,16 @@ import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import dev.enro.animation.NavigationAnimationOverrideBuilder
-import dev.enro.core.*
+import dev.enro.core.NavigationDirection
+import dev.enro.core.NavigationInstruction
+import dev.enro.core.NavigationKey
 import dev.enro.core.container.EmptyBehavior
 import dev.enro.core.container.NavigationContainerProperty
+import dev.enro.core.container.NavigationInstructionFilter
+import dev.enro.core.container.acceptAll
 import dev.enro.core.container.backstackOfNotNull
 import dev.enro.core.controller.interceptor.builder.NavigationInterceptorBuilder
+import dev.enro.core.navigationContext
 
 
 public fun FragmentActivity.navigationContainer(
@@ -17,14 +22,14 @@ public fun FragmentActivity.navigationContainer(
     emptyBehavior: EmptyBehavior = EmptyBehavior.AllowEmpty,
     interceptor: NavigationInterceptorBuilder.() -> Unit = {},
     animations: NavigationAnimationOverrideBuilder.() -> Unit = {},
-    accept: (NavigationKey) -> Boolean = { true },
+    filter: NavigationInstructionFilter = acceptAll(),
 ): NavigationContainerProperty<FragmentNavigationContainer> = navigationContainer(
     containerId = containerId,
     rootInstruction = { root()?.let { NavigationInstruction.Push(it) } },
     emptyBehavior = emptyBehavior,
     interceptor = interceptor,
     animations = animations,
-    accept = accept,
+    filter = filter,
 )
 
 @JvmName("navigationContainerFromInstruction")
@@ -34,14 +39,14 @@ public fun FragmentActivity.navigationContainer(
     emptyBehavior: EmptyBehavior = EmptyBehavior.AllowEmpty,
     interceptor: NavigationInterceptorBuilder.() -> Unit = {},
     animations: NavigationAnimationOverrideBuilder.() -> Unit = {},
-    accept: (NavigationKey) -> Boolean = { true },
+    filter: NavigationInstructionFilter = acceptAll(),
 ): NavigationContainerProperty<FragmentNavigationContainer> = NavigationContainerProperty(
     lifecycleOwner = this,
     navigationContainerProducer = {
         FragmentNavigationContainer(
             containerId = containerId,
             parentContext = navigationContext,
-            accept = accept,
+            filter = filter,
             emptyBehavior = emptyBehavior,
             interceptor = interceptor,
             animations = animations,
@@ -56,14 +61,14 @@ public fun Fragment.navigationContainer(
     emptyBehavior: EmptyBehavior = EmptyBehavior.AllowEmpty,
     interceptor: NavigationInterceptorBuilder.() -> Unit = {},
     animations: NavigationAnimationOverrideBuilder.() -> Unit = {},
-    accept: (NavigationKey) -> Boolean = { true },
+    filter: NavigationInstructionFilter = acceptAll(),
 ): NavigationContainerProperty<FragmentNavigationContainer> = navigationContainer(
     containerId = containerId,
     rootInstruction = { root()?.let { NavigationInstruction.Push(it) } },
     emptyBehavior = emptyBehavior,
     interceptor = interceptor,
     animations = animations,
-    accept = accept,
+    filter = filter,
 )
 
 @JvmName("navigationContainerFromInstruction")
@@ -73,14 +78,14 @@ public fun Fragment.navigationContainer(
     emptyBehavior: EmptyBehavior = EmptyBehavior.AllowEmpty,
     interceptor: NavigationInterceptorBuilder.() -> Unit = {},
     animations: NavigationAnimationOverrideBuilder.() -> Unit = {},
-    accept: (NavigationKey) -> Boolean = { true },
+    filter: NavigationInstructionFilter = acceptAll(),
 ): NavigationContainerProperty<FragmentNavigationContainer> = NavigationContainerProperty(
     lifecycleOwner = this,
     navigationContainerProducer = {
         FragmentNavigationContainer(
             containerId = containerId,
             parentContext = navigationContext,
-            accept = accept,
+            filter = filter,
             emptyBehavior = emptyBehavior,
             interceptor = interceptor,
             animations = animations,