+++ enro/src/androidTest/java/dev/enro/TestViews.kt
@@ -35,6 +35,7 @@ import dev.enro.core.compose.EnroContainer
 import dev.enro.core.compose.navigationHandle
 import dev.enro.core.compose.rememberNavigationContainer
 import dev.enro.core.container.EmptyBehavior
+import dev.enro.core.container.acceptKey
 import dev.enro.core.getNavigationHandle
 
 abstract class TestActivity : AppCompatActivity() {
@@ -232,12 +233,12 @@ fun TestComposable(
     secondaryContainerAccepts: (NavigationKey) -> Boolean = { false }
 ) {
     val primaryContainer = rememberNavigationContainer(
-        accept = primaryContainerAccepts,
+        filter = acceptKey(primaryContainerAccepts),
         emptyBehavior = EmptyBehavior.AllowEmpty,
     )
 
     val secondaryContainer = rememberNavigationContainer(
-        accept = secondaryContainerAccepts,
+        filter = acceptKey(secondaryContainerAccepts),
         emptyBehavior = EmptyBehavior.AllowEmpty,
     )